requirements.txt
aio-pika==9.1.2
aiogram==2.25.1
aiohttp==3.8.4
aioredis==2.0.1
aiormq==6.7.6
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
anyio==3.7.0
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
Babel==2.9.1
backcall==0.2.0
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dataclasses-json==0.5.8
decorator==5.1.1
docker==6.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
executing==1.2.0
frozenlist==1.3.3
fsspec==2023.6.0
fuzzywuzzy==0.18.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
ipython==8.14.0
jedi==0.18.2
kombu==5.3.0
langchain==0.0.206
langchainplus-sdk==0.0.16
Levenshtein==0.21.1
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mirakuru==2.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentele==1.15.1
packaging==23.1
pamqp==3.2.1
pandas==2.0.2
parso==0.8.3
passlib==1.7.4
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.2
pika-stubs==0.1.3
pluggy==1.2.0
port-for==0.7.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg==3.1.9
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.5.0
pydantic==1.10.9
Pygments==2.15.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pytest==7.3.2
pytest-asyncio==0.21.0
pytest-env==0.8.2
pytest-postgresql==5.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-Levenshtein==0.21.1
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.1.1
redis==4.5.5
regex==2023.6.3
requests==2.31.0
rsa==4.9
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
SQLAlchemy-Utils==0.41.1
stack-data==0.6.2
structlog==23.1.0
Telethon==1.28.5
tenacity==8.2.2
TgCrypto==1.2.5
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.1
yarl==1.9.2
